+++ Code/lib/Xila/include/Software/Clock.hpp
@@ -122,7 +122,7 @@ private:
     void Add_Alarm();
     void Delete_Alarm();
 
-    void Set_Variable(Xila_Class::Adress, uint8_t, const void *);
+    void Set_Variable(Xila_Class::Address, uint8_t, const void *);
 
     uint8_t Selected_Alarm;