+++ Code/lib/Xila/src/Core/Modules/Dialog.cpp
@@ -36,9 +36,9 @@ Xila_Class::Event Xila_Class::Dialog_Class::Keyboard(char *String, size_t Size, 
   Xila_Class::Page Page = Xila.Display.Get_Current_Page();
   // -- Initalize shell
   Xila.Software_Management.Shell_Send_Instruction(Dialog_Keyboard);
-  Xila.Software_Management.Shell_Set_Variable(Adress('S', 't'), Variable_String, String);
-  Xila.Software_Management.Shell_Set_Variable(Adress('S', 'i'), Variable_Long, &Size);
-  Xila.Software_Management.Shell_Set_Variable(Adress('M', 'a'), Variable_Long, &Masked_Input);
+  Xila.Software_Management.Shell_Set_Variable(Address('S', 't'), Variable_String, String);
+  Xila.Software_Management.Shell_Set_Variable(Address('S', 'i'), Variable_Long, &Size);
+  Xila.Software_Management.Shell_Set_Variable(Address('M', 'a'), Variable_Long, &Masked_Input);
   Xila.Software_Management.Maximize(Shell_Handle);
   // -- Tasks suspended here
   while (Keyboard_State == Xila.None)
@@ -70,7 +70,7 @@ Xila_Class::Event Xila_Class::Dialog_Class::Keypad(float &Number)
   Xila_Class::Page Page = Xila.Display.Get_Current_Page();
   // -- Initalize shell
   Xila.Software_Management.Shell_Send_Instruction(Dialog_Keypad);
-  Xila.Software_Management.Shell_Set_Variable(Adress('N', 'u'), Variable_Long, &Number);
+  Xila.Software_Management.Shell_Set_Variable(Address('N', 'u'), Variable_Long, &Number);
   Xila.Software_Management.Maximize(Shell_Handle);
   // -- Tasks suspended here
   while (Keypad_State == Xila.None)
@@ -102,7 +102,7 @@ Xila_Class::Event Xila_Class::Dialog_Class::Color_Picker(uint16_t &Color)
   Xila_Class::Page Page = Xila.Display.Get_Current_Page();
   // -- Initalize shell
   Xila.Software_Management.Shell_Send_Instruction(Dialog_Color_Picker);
-  Xila.Software_Management.Shell_Set_Variable(Adress('C', 'P'), Pointer, &Color);
+  Xila.Software_Management.Shell_Set_Variable(Address('C', 'P'), Pointer, &Color);
   Xila.Software_Management.Maximize(Shell_Handle);
   // -- Task suspended here
   while (Color_Picker_State == Xila.None)
@@ -139,12 +139,12 @@ Xila_Class::Event Xila_Class::Dialog_Class::Event(const __FlashStringHelper *Mes
   Xila_Class::Page Page = Xila.Display.Get_Current_Page();
   // -- Initalize shell
   Xila.Software_Management.Shell_Send_Instruction(Dialog_Event);
-  Xila.Software_Management.Shell_Set_Variable(Adress('M', 'e'), Variable_String, Message);
-  Xila.Software_Management.Shell_Set_Variable(Adress('T', 'y'), Variable_Long, &Type);
-  Xila.Software_Management.Shell_Set_Variable(Adress('M', 'o'), Variable_Long, &Mode);
-  Xila.Software_Management.Shell_Set_Variable(Adress('B', '1'), Variable_String, Button_Text_1);
-  Xila.Software_Management.Shell_Set_Variable(Adress('B', '2'), Variable_String, Button_Text_2);
-  Xila.Software_Management.Shell_Set_Variable(Adress('B', '3'), Variable_String, Button_Text_3);
+  Xila.Software_Management.Shell_Set_Variable(Address('M', 'e'), Variable_String, Message);
+  Xila.Software_Management.Shell_Set_Variable(Address('T', 'y'), Variable_Long, &Type);
+  Xila.Software_Management.Shell_Set_Variable(Address('M', 'o'), Variable_Long, &Mode);
+  Xila.Software_Management.Shell_Set_Variable(Address('B', '1'), Variable_String, Button_Text_1);
+  Xila.Software_Management.Shell_Set_Variable(Address('B', '2'), Variable_String, Button_Text_2);
+  Xila.Software_Management.Shell_Set_Variable(Address('B', '3'), Variable_String, Button_Text_3);
   Xila.Software_Management.Maximize(Shell_Handle);
   // -- Tasks suspended here
   while (Event_State == Xila.None)
@@ -181,12 +181,12 @@ Xila_Class::Event Xila_Class::Dialog_Class::Event(const char *Message, uint8_t T
   Xila_Class::Page Page = Xila.Display.Get_Current_Page();
   // -- Initalize shell variable
   Xila.Software_Management.Shell_Send_Instruction(Dialog_Event);
-  Xila.Software_Management.Shell_Set_Variable(Adress('M', 'e'), Variable_String, Message);
-  Xila.Software_Management.Shell_Set_Variable(Adress('T', 'y'), Variable_Long, &Type);
-  Xila.Software_Management.Shell_Set_Variable(Adress('M', 'o'), Variable_Long, &Mode);
-  Xila.Software_Management.Shell_Set_Variable(Adress('B', '1'), Variable_String, Button_Text_1);
-  Xila.Software_Management.Shell_Set_Variable(Adress('B', '2'), Variable_String, Button_Text_2);
-  Xila.Software_Management.Shell_Set_Variable(Adress('B', '3'), Variable_String, Button_Text_3);
+  Xila.Software_Management.Shell_Set_Variable(Address('M', 'e'), Variable_String, Message);
+  Xila.Software_Management.Shell_Set_Variable(Address('T', 'y'), Variable_Long, &Type);
+  Xila.Software_Management.Shell_Set_Variable(Address('M', 'o'), Variable_Long, &Mode);
+  Xila.Software_Management.Shell_Set_Variable(Address('B', '1'), Variable_String, Button_Text_1);
+  Xila.Software_Management.Shell_Set_Variable(Address('B', '2'), Variable_String, Button_Text_2);
+  Xila.Software_Management.Shell_Set_Variable(Address('B', '3'), Variable_String, Button_Text_3);
   Xila.Software_Management.Maximize(Shell_Handle);
   // -- Tasks suspended here
   while (Event_State == Xila.None)
@@ -221,12 +221,12 @@ void Xila_Class::Dialog_Class::Load(const __FlashStringHelper *Header, const __F
   Xila_Class::Page Page = Xila.Display.Get_Current_Page();
   // -- Initalize shell
   Xila.Software_Management.Shell_Send_Instruction(Dialog_Load);
-  Xila.Software_Management.Shell_Set_Variable(Adress('H', 'e'), Variable_String, Header);
-  Xila.Software_Management.Shell_Set_Variable(Adress('M', 'e'), Variable_String, Message);
-  Xila.Software_Management.Shell_Set_Variable(Adress('M', 'o'), Variable_Long, &Mode);
-  Xila.Software_Management.Shell_Set_Variable(Adress('D', 'u'), Variable_Long, &Duration);
-  Xila.Software_Management.Shell_Set_Variable(Adress('S', 'o'), Variable_Long, Xila.Software_Management.Openned[0]);
-  Xila.Software_Management.Shell_Set_Variable(Adress('P', 'a'), Variable_Long, &Page);
+  Xila.Software_Management.Shell_Set_Variable(Address('H', 'e'), Variable_String, Header);
+  Xila.Software_Management.Shell_Set_Variable(Address('M', 'e'), Variable_String, Message);
+  Xila.Software_Management.Shell_Set_Variable(Address('M', 'o'), Variable_Long, &Mode);
+  Xila.Software_Management.Shell_Set_Variable(Address('D', 'u'), Variable_Long, &Duration);
+  Xila.Software_Management.Shell_Set_Variable(Address('S', 'o'), Variable_Long, Xila.Software_Management.Openned[0]);
+  Xila.Software_Management.Shell_Set_Variable(Address('P', 'a'), Variable_Long, &Page);
   Xila.Software_Management.Openned[0]->Set_Watchdog_Timeout(30000);
   Xila.Software_Management.Maximize(Shell_Handle);
   // -- Tasks suspended here
@@ -252,12 +252,12 @@ void Xila_Class::Dialog_Class::Load(const char *Header, const char *Message, uin
   Xila_Class::Page Page = Xila.Display.Get_Current_Page();
   // -- Initalize shell
   Xila.Software_Management.Shell_Send_Instruction(Dialog_Load);
-  Xila.Software_Management.Shell_Set_Variable(Adress('H', 'e'), Variable_String, Header);
-  Xila.Software_Management.Shell_Set_Variable(Adress('M', 'e'), Variable_String, Message);
-  Xila.Software_Management.Shell_Set_Variable(Adress('M', 'o'), Variable_Long, &Mode);
-  Xila.Software_Management.Shell_Set_Variable(Adress('D', 'u'), Variable_Long, &Duration);
-  Xila.Software_Management.Shell_Set_Variable(Adress('S', 'o'), Variable_Long, Xila.Software_Management.Openned[0]);
-  Xila.Software_Management.Shell_Set_Variable(Adress('P', 'a'), Variable_Long, &Page);
+  Xila.Software_Management.Shell_Set_Variable(Address('H', 'e'), Variable_String, Header);
+  Xila.Software_Management.Shell_Set_Variable(Address('M', 'e'), Variable_String, Message);
+  Xila.Software_Management.Shell_Set_Variable(Address('M', 'o'), Variable_Long, &Mode);
+  Xila.Software_Management.Shell_Set_Variable(Address('D', 'u'), Variable_Long, &Duration);
+  Xila.Software_Management.Shell_Set_Variable(Address('S', 'o'), Variable_Long, Xila.Software_Management.Openned[0]);
+  Xila.Software_Management.Shell_Set_Variable(Address('P', 'a'), Variable_Long, &Page);
   Xila.Software_Management.Openned[0]->Set_Watchdog_Timeout(30000);
   Xila.Software_Management.Maximize(Shell_Handle);
   // -- Tasks suspended here
@@ -270,7 +270,7 @@ void Xila_Class::Dialog_Class::Load(const char *Header, const char *Message, uin
 void Xila_Class::Dialog_Class::Close_Load()
 {
   uint32_t Duration = Xila.Time.Milliseconds();
-  Xila.Software_Management.Shell_Set_Variable(Adress('l', 'D'), Variable_Long, &Duration);
+  Xila.Software_Management.Shell_Set_Variable(Address('l', 'D'), Variable_Long, &Duration);
 }
 
 ///
@@ -345,7 +345,7 @@ Xila_Class::Event Xila_Class::Dialog_Class::Open_File(File &File_To_Open)
   Xila_Class::Page Page = Xila.Display.Get_Current_Page();
   // -- Initalize shell
     Xila.Software_Management.Shell_Send_Instruction(Dialog_Open_File);
-  Xila.Software_Management.Shell_Set_Variable(Adress('I', 'P'), Pointer, &File_To_Open);
+  Xila.Software_Management.Shell_Set_Variable(Address('I', 'P'), Pointer, &File_To_Open);
   Xila.Software_Management.Maximize(Shell_Handle);
   // -- Tasks suspended here
   while (File_Manager_State == Xila.None)
@@ -377,7 +377,7 @@ Xila_Class::Event Xila_Class::Dialog_Class::Save_File(File &File_To_Save)
   Xila_Class::Page Page = Xila.Display.Get_Current_Page();
   // -- Initalize shell
     Xila.Software_Management.Shell_Send_Instruction(Dialog_Save_File);
-  Xila.Software_Management.Shell_Set_Variable(Adress('I', 'P'), Pointer, &File_To_Save);
+  Xila.Software_Management.Shell_Set_Variable(Address('I', 'P'), Pointer, &File_To_Save);
   Xila.Software_Management.Maximize(Shell_Handle);
   // -- Tasks suspended here
   while (File_Manager_State == Xila.None)
@@ -409,7 +409,7 @@ Xila_Class::Event Xila_Class::Dialog_Class::Open_Folder(File &Folder_To_Open)
   Xila_Class::Page Page = Xila.Display.Get_Current_Page();
   // -- Initalize shell variable
     Xila.Software_Management.Shell_Send_Instruction(Dialog_Open_Folder);
-  Xila.Software_Management.Shell_Set_Variable(Adress('I', 'P'), Pointer, &Folder_To_Open);
+  Xila.Software_Management.Shell_Set_Variable(Address('I', 'P'), Pointer, &Folder_To_Open);
   Xila.Software_Management.Maximize(Shell_Handle);
   // -- Tasks suspended here
   while (File_Manager_State == Xila.None)