+++ Changelog.md
@@ -33,7 +33,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - Fix deprecated function in Periodic (with "as.<const char*()>" instead of "as.<char*>()").
 - Update sound driver.
 - Change Internet Browser home page.
-- Change variable adress from 8 bits to 16 bits.
+- Change variable Address from 8 bits to 16 bits.
 - Change Xila.Dialog.Dialog_Open_Folder(File&) to Xila.Dialog.Open_Folder(File&).
 - Header refresh is handle by Shell, not by the Core anymore.
 - How core and Shell and Core interracts for Dialogs.